+++ elena/constants.py
@@ -1,4 +1,0 @@
-import pkg_resources  # part of setuptools
-
-VERSION = pkg_resources.require("elena")[0].version
-