+++ elena/domain/ports/exchange_manager.py
@@ -26,7 +26,7 @@ class ExchangeManager(Protocol):
         :param time_frame: time frame to read
         :return: the market candles
         """
-        pass
+        ...
 
     def read_order_book(
             self,
@@ -39,7 +39,9 @@ class ExchangeManager(Protocol):
         :param pair: trading pair to read
         :return: the current order book
         """
-        pass
+        ...
+
+    # TODO read all our orders
 
     def get_balance(
             self,
@@ -50,13 +52,13 @@ class ExchangeManager(Protocol):
         :param exchange: exchange where to read market data
         :return: The balance
         """
-        pass
+        ...
 
     def place_order(
             self,
             exchange: Exchange,
             bot_config: BotConfig,
-            type: OrderType,
+            order_type: OrderType,
             side: OrderSide,
             amount: float,
             price: Optional[float] = None
@@ -65,40 +67,40 @@ class ExchangeManager(Protocol):
         Places an order to an Exchange
         :param exchange: exchange where to read market data
         :param bot_config: the current bot configuration
-        :param type: the type of your order
+        :param order_type: the type of your order
         :param side: the direction of your order
         :param amount: how much of currency you want to trade
         :param price: the price at which the order is to be fulfilled (ignored in market orders)
         :return: the placed Order, error if any
         """
-        pass
+        ...
 
     def cancel_order(
             self,
             exchange: Exchange,
             bot_config: BotConfig,
-            id: str
+            order_id: str
     ):
         """
         Cancels an order on a Exchange
         :param exchange: exchange where to read market data
         :param bot_config: the current bot configuration
-        :param id: the order id to cancel_order
+        :param order_id: the order id to cancel_order
         :return: error if any
         """
-        pass
+        ...
 
     def fetch_order(
             self,
             exchange: Exchange,
             bot_config: BotConfig,
-            id: str
+            order_id: str
     ) -> Order:
         """
         Retrieves an order from Exchange
         :param exchange: exchange where to read market data
         :param bot_config: the current bot configuration
-        :param id: the order id to retrieve
+        :param order_id: the order id to retrieve
         :return: the list of orders, error if any
         """
-        pass
+        ...