+++ elena/entrypoint/local/entrypoint.py
@@ -1,6 +1,6 @@
+from elena import __version__ as version
+from elena.adapters.config.local_config_reader import LocalConfigReader
 from elena.config import dependency_injection
-from elena.constants import VERSION
-from elena.entrypoint.local.local_config_reader import LocalConfigReader
 
 
 def main():
@@ -8,7 +8,7 @@ def main():
     container = dependency_injection.get_container(config)
     container.wire(modules=[__name__])
     elena = container.elena()
-    print(f"Starting Elena v{VERSION} from CLI")
+    container.logger().info("Starting Elena v%s from CLI", version)
     elena.run()